CHROMIUM: mmc: sdhci-tegra: Set MMC_CAP_WAIT_WHILE_BUSY
Although MMC_CAP_WAIT_WHILE_BUSY is not well-documented, it appears that
it should be set when the host-controller is capable of determining
whether the card is busy or not and polls this state before attempting
to submit a request.  This likely applies to all SDHCI hosts, but we set
it now only for Tegra to minimize the possible impact.

BUG=b:25821594
TEST=Suspend/resume on Smaug.

Change-Id: Ie15e6cb4ca1ea2669e58860fe5925c267b973d08
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/316469
Reviewed-by: Benson Leung <bleung@chromium.org>